feat: repository finished

+++ src/modules/workers/services.js
@@ -2,7 +2,14 @@ const workersSchema = require('./model')
 
 const getAllWorkers = async () => await workersSchema.find({})
 
-const getOneWorker = async (id) => await workersSchema.findOne({ _id: id })
+const getOneWorker = async (id) => {
+    const worker = await workersSchema.findOne({ _id: id })
+    if (!worker) {
+        return 'Not found'
+    }
+
+    return worker
+}
 
 const addWorker = async (data) => {
     const newWorker = new workersSchema({
@@ -15,6 +22,10 @@ const addWorker = async (data) => {
 }
 
 const updateWorker = async (id, data) => {
+    const worker = await workersSchema.findOne({ _id: id })
+    if (!worker) {
+        return 'Not found'
+    }
     await workersSchema.updateOne({ _id: id }, { $set: data }).exec((err) => {
         if (err) {
             return err
@@ -25,6 +36,10 @@ const updateWorker = async (id, data) => {
 }
 
 const deleteWorker = async (id) => {
+    const worker = await workersSchema.findOne({ _id: id })
+    if (!worker) {
+        return 'Not found'
+    }
     await workersSchema.deleteOne({ _id: id })
 
     return 'Deleted'
